hardware/device-hub/src/serial_controller.py
# Serial Controller for Arduino Communication
# TODO: Implement USB Serial communication

import serial
import time

class SerialController:
    def __init__(self, port: str = "/dev/ttyUSB0", baudrate: int = 9600):
        self.port = port
        self.baudrate = baudrate
        self.connection = None
        
    def connect(self):
        """Connect to Arduino via USB Serial"""
        # TODO: Initialize serial connection
        pass
    
    def send_command(self, command: str):
        """Send command to Arduino"""
        # TODO: Send serial command
        pass
    
    def read_response(self) -> str:
        """Read response from Arduino"""
        # TODO: Read serial response
        return ""
    
    def disconnect(self):
        """Close serial connection"""
        # TODO: Close connection
        pass