hardware/device-hub/requirements.txt
# Python Dependencies for Device Hub
# TODO: Add production dependencies

websockets==11.0.3
pyserial==3.5
asyncio
RPi.GPIO==0.7.1
psutil==5.9.5